# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parallel AI Coding Orchestrator API",
    "description": "\n## Week 2 MVP - Manager AI Integration API\n\nEnterprise-grade REST API for orchestrating and monitoring parallel AI coding workflows.\n\n### Key Features\n- **Supervisor API**: Monitor and control Claude Code worker instances\n- **Resource Management**: Hierarchical resource allocation with depth-based quotas\n- **Job Orchestration**: Manage job lifecycle with state machine validation\n- **Authentication**: JWT-based authentication with scope-based authorization\n\n### Architecture\n- **Database**: SQLite (development) / PostgreSQL (production)\n- **Authentication**: JWT Bearer tokens with scopes\n- **State Management**: Validated state machines for workers and jobs\n- **Testing**: 44 integration tests with 29% code coverage\n\n### Quick Start\n1. Generate test token: Use `/api/auth/token` endpoint\n2. Access protected endpoints with `Authorization: Bearer <token>` header\n3. Explore interactive docs at `/docs` or `/redoc`\n\n### Support\n- **GitHub**: [parallel-coding](https://github.com/your-org/parallel-coding)\n- **Documentation**: [Wiki](https://github.com/your-org/parallel-coding/wiki)\n",
    "contact": {
      "name": "Parallel AI Development Team",
      "email": "dev@parallel-ai.example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "2.0.0-week2-mvp"
  },
  "paths": {
...
    "/api/v1/workers/{worker_id}": {
      "get": {
        "summary": "Get Worker Info",
        "description": "Get information about a specific worker.\n\nArgs:\n    worker_id: Worker identifier (e.g., \"worker_001\")\n\nReturns:\n    Worker details including dialogue status\n\nRaises:\n    HTTPException: If worker not found",
        "operationId": "get_worker_info_api_v1_workers__worker_id__get",
        "parameters": [
          {
            "name": "worker_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Worker Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}